Added simple database object to model.
This class extends the SQLiteConnection class, and is used to interact with
the database.  Its constructor checks if a file exists at the specified path
and, if one does not, creates a database there.  We shall see if this is
generic enough to be in the model and not in the platform-specific code.