Changed reference to Phonograph to a linked project.
Using a project reference did not work, as the two project are configured to
compile to two different frameworks (Android and .NET).  Linking the
Phonograph project to Phonograph.Android makes the files in Phonograph appear
to be part of Phonograph.Android, which allows them to be used and compiled
in.