Added a "view" of the plays table.
This is not a true view, as it is not in the database itself.  But it is a set
of columns that are used after joins on all the tables the plays table
references.  This needs to be done because the SQLite library does not support
joins via linq to SQL, so a raw query must be used.