Store database in a permanent location.
The "Personal" special folder was being deleted each time the application was
uninstalled, and thus the plays were lost each time.  Instead, store the
database in a more generic location, which should not be regularly deleted.

Hopefully this location will not cause issues on some devices, but it might.
I can come up with a better way to reference that location later.