Added broadcast receiver for Pocket Casts.
Unfortunately, as indicated in the comments, this broadcast receiver is
never used, as it uses the default intents.  It also does not properly
set the "playing" extra, so plays are not recorded.  I am uncomfortable
with the idea of extra engineering to handle Pocket Casts specifically
in the generic broadcast receiver.